requirements.txt
numpy~=1.19.5
tensorflow~=2.5.0
fer~=20.1.3
setuptools~=56.2.0
opencv-python~=4.5.2.52
pandas~=1.2.4
matplotlib~=3.4.1
scikit-learn~=0.24.2
pytest~=6.2.4
mock~=4.0.3
Pillow~=8.2.0
mkdocs~=3.0.0